Move component PositiveRateCard from 'components/cards' to 'components/index/CardsMonitoring'
