+++ components/index/CardsMonitoring/PositiveRateCard.vue
